Preferences Fixes - Moved all preferences to app.config - Renamed prefernces service to  appconfig - Language setting now saved in appconfig -  Need to get theme to save on change
