+++ src/Nisshi/ClientApp/src/app/core/config/app.config.ts
@@ -13,6 +13,35 @@ export interface AppConfig
     layout: Layout;
     scheme: Scheme;
     theme: Theme;
+    language: string;
+    logbookOptions: LogbookOptions;
+}
+
+/**
+ * LogbookConfig interface. Contains configuration for columns to display on the
+ * logbook view
+ */
+export interface LogbookOptions
+{
+    showTailNumber: boolean;
+    showTypeName: boolean;
+    showApproaches: boolean;
+    showLandings: boolean;
+    showNightLandings: boolean;
+    showFullStopLandings: boolean;
+    showPIC: boolean;
+    showSIC: boolean;
+    showCrossCountry: boolean;
+    showNight: boolean;
+    showMultiEngine: boolean;
+    showTurbine: boolean;
+    showSimulatedInstrument: boolean;
+    showIMC: boolean;
+    showDualReceived: boolean;
+    showDualGiven: boolean;
+    showGroundSim: boolean;
+    showTotalTime: boolean;
+    showComments: boolean;
 }
 
 /**
@@ -26,5 +55,27 @@ export interface AppConfig
 export const appConfig: AppConfig = {
     layout: 'classy',
     scheme: 'light',
-    theme : 'default'
+    theme : 'default',
+    language: 'en',
+        logbookOptions: {
+            showTailNumber: true,
+            showTypeName: true,
+            showApproaches: true,
+            showLandings: true,
+            showNightLandings: true,
+            showFullStopLandings: true,
+            showCrossCountry: true,
+            showNight: true,
+            showPIC: true,
+            showSIC: true,
+            showMultiEngine: true,
+            showTurbine: true,
+            showSimulatedInstrument: true,
+            showIMC: true,
+            showDualReceived: true,
+            showDualGiven: true,
+            showGroundSim: true,
+            showTotalTime: true,
+            showComments: true,
+        }
 };
