--- src/Nisshi/ClientApp/src/app/core/preferences/preferences.types.ts
@@ -1,31 +1,0 @@
-export interface Preferences
-{
-    layout: string;
-    scheme: string;
-    theme: string;
-    language: string;
-    logbookOptions: LogbookOptions;
-}
-
-export interface LogbookOptions
-{
-    showTailNumber: boolean;
-    showTypeName: boolean;
-    showApproaches: boolean;
-    showLandings: boolean;
-    showNightLandings: boolean;
-    showFullStopLandings: boolean;
-    showPIC: boolean;
-    showSIC: boolean;
-    showCrossCountry: boolean;
-    showNight: boolean;
-    showMultiEngine: boolean;
-    showTurbine: boolean;
-    showSimulatedInstrument: boolean;
-    showIMC: boolean;
-    showDualReceived: boolean;
-    showDualGiven: boolean;
-    showGroundSim: boolean;
-    showTotalTime: boolean;
-    showComments: boolean;
-}
