+++ src/Nisshi/ClientApp/src/app/core/user/user.types.ts
@@ -1,4 +1,3 @@
-import { Preferences } from "../preferences/preferences.types";
 
 export interface User
 {
