+++ src/Nisshi/ClientApp/src/app/layout/common/settings/settings.component.ts
@@ -4,8 +4,9 @@ import { Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
 import { FuseConfigService } from '@fuse/services/config';
 import { FuseTailwindService } from '@fuse/services/tailwind';
-import { AppConfig, Scheme, Theme } from 'app/core/config/app.config';
+import { appConfig, AppConfig, Scheme, Theme } from 'app/core/config/app.config';
 import { Layout } from 'app/layout/layout.types';
+import { AppConfigService } from 'app/core/config/appconfig.service';
 
 @Component({
     selector     : 'settings',
@@ -37,7 +38,8 @@ export class SettingsComponent implements OnInit, OnDestroy
     constructor(
         private _router: Router,
         private _fuseConfigService: FuseConfigService,
-        private _fuseTailwindService: FuseTailwindService
+        private _fuseTailwindService: FuseTailwindService,
+        private _appConfigService: AppConfigService
     )
     {
     }
@@ -66,6 +68,8 @@ export class SettingsComponent implements OnInit, OnDestroy
                 // Store the config
                 this.config = config;
             });
+
+        //this.config = this._appConfigService.appConfig$;
     }
 
     /**
@@ -99,6 +103,8 @@ export class SettingsComponent implements OnInit, OnDestroy
 
             // Set the config
             this._fuseConfigService.config = {layout};
+
+            this._appConfigService.appConfig = this._fuseConfigService.config;
         });
     }
 
@@ -110,6 +116,8 @@ export class SettingsComponent implements OnInit, OnDestroy
     setScheme(scheme: Scheme): void
     {
         this._fuseConfigService.config = {scheme};
+
+        this._appConfigService.appConfig = this._fuseConfigService.config;
     }
 
     /**
@@ -120,5 +128,7 @@ export class SettingsComponent implements OnInit, OnDestroy
     setTheme(theme: Theme): void
     {
         this._fuseConfigService.config = {theme};
+
+        this._appConfigService.appConfig = this._fuseConfigService.config;
     }
 }
