+++ src/Nisshi/ClientApp/src/app/modules/logbook/form/logbook-form.component.html
@@ -111,7 +111,7 @@
                     </p>
                 </div>
                 <div class="flex-auto min-w-64">
-                    <div class="flex" *ngIf="this.preferences.logbookOptions.showComments">
+                    <div class="flex" *ngIf="this.appConfig.logbookOptions.showComments">
                         <mat-form-field class="fuse-mat-textarea flex-auto">
                             <mat-label>Comments</mat-label>
                             <textarea
@@ -128,7 +128,7 @@
                     </div>
                 </div>
                 <div class="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 min-w-64 justify-between">
-                    <div class="flex mx-2" *ngIf="this.preferences.logbookOptions.showApproaches">
+                    <div class="flex mx-2" *ngIf="this.appConfig.logbookOptions.showApproaches">
                         <mat-form-field class="flex-auto">
                             <mat-label>Number of Instrument Approaches</mat-label>
                             <input
@@ -148,7 +148,7 @@
                             </mat-error>
                         </mat-form-field>
                     </div>
-                    <div class="flex mx-2" *ngIf="this.preferences.logbookOptions.showLandings">
+                    <div class="flex mx-2" *ngIf="this.appConfig.logbookOptions.showLandings">
                         <mat-form-field class="flex-auto">
                             <mat-label>Number of Landings</mat-label>
                             <input
@@ -168,7 +168,7 @@
                             </mat-error>
                         </mat-form-field>
                     </div>
-                    <div class="flex mx-2" *ngIf="this.preferences.logbookOptions.showNightLandings">
+                    <div class="flex mx-2" *ngIf="this.appConfig.logbookOptions.showNightLandings">
                         <mat-form-field class="flex-auto">
                             <mat-label>Number of Night Landings</mat-label>
                             <input
@@ -188,7 +188,7 @@
                             </mat-error>
                         </mat-form-field>
                     </div>
-                    <div class="flex mx-2" *ngIf="this.preferences.logbookOptions.showFullStopLandings">
+                    <div class="flex mx-2" *ngIf="this.appConfig.logbookOptions.showFullStopLandings">
                         <mat-form-field class="flex-auto">
                             <mat-label>Number of Full-Stop Landings</mat-label>
                             <input
@@ -208,7 +208,7 @@
                             </mat-error>
                         </mat-form-field>
                     </div>
-                    <div class="flex mx-2" *ngIf="this.preferences.logbookOptions.showCrossCountry">
+                    <div class="flex mx-2" *ngIf="this.appConfig.logbookOptions.showCrossCountry">
                         <mat-form-field class="flex-auto">
                             <mat-label>Cross Country</mat-label>
                             <input
@@ -228,7 +228,7 @@
                             </mat-error>
                         </mat-form-field>
                     </div>
-                    <div class="flex mx-2" *ngIf="this.preferences.logbookOptions.showMultiEngine">
+                    <div class="flex mx-2" *ngIf="this.appConfig.logbookOptions.showMultiEngine">
                         <mat-form-field class="flex-auto">
                             <mat-label>Multi Engine</mat-label>
                             <input
@@ -248,7 +248,7 @@
                             </mat-error>
                         </mat-form-field>
                     </div>
-                    <div class="flex mx-2" *ngIf="this.preferences.logbookOptions.showTurbine">
+                    <div class="flex mx-2" *ngIf="this.appConfig.logbookOptions.showTurbine">
                         <mat-form-field class="flex-auto">
                             <mat-label>Turbine</mat-label>
                             <input
@@ -268,7 +268,7 @@
                             </mat-error>
                         </mat-form-field>
                     </div>
-                    <div class="flex mx-2" *ngIf="this.preferences.logbookOptions.showNight">
+                    <div class="flex mx-2" *ngIf="this.appConfig.logbookOptions.showNight">
                         <mat-form-field class="flex-auto">
                             <mat-label>Night</mat-label>
                             <input
@@ -288,7 +288,7 @@
                             </mat-error>
                         </mat-form-field>
                     </div>
-                    <div class="flex mx-2" *ngIf="this.preferences.logbookOptions.showIMC">
+                    <div class="flex mx-2" *ngIf="this.appConfig.logbookOptions.showIMC">
                         <mat-form-field class="flex-auto">
                             <mat-label>Actual IMC</mat-label>
                             <input
@@ -308,7 +308,7 @@
                             </mat-error>
                         </mat-form-field>
                     </div>
-                    <div class="flex mx-2" *ngIf="this.preferences.logbookOptions.showSimulatedInstrument">
+                    <div class="flex mx-2" *ngIf="this.appConfig.logbookOptions.showSimulatedInstrument">
                         <mat-form-field class="flex-auto">
                             <mat-label>Simulated IMC</mat-label>
                             <input
@@ -328,7 +328,7 @@
                             </mat-error>
                         </mat-form-field>
                     </div>
-                    <div class="flex mx-2" *ngIf="this.preferences.logbookOptions.showDualReceived">
+                    <div class="flex mx-2" *ngIf="this.appConfig.logbookOptions.showDualReceived">
                         <mat-form-field class="flex-auto">
                             <mat-label>Dual Received</mat-label>
                             <input
@@ -348,7 +348,7 @@
                             </mat-error>
                         </mat-form-field>
                     </div>
-                    <div class="flex mx-2" *ngIf="this.preferences.logbookOptions.showDualGiven">
+                    <div class="flex mx-2" *ngIf="this.appConfig.logbookOptions.showDualGiven">
                         <mat-form-field class="flex-auto">
                             <mat-label>Dual Given</mat-label>
                             <input
@@ -368,7 +368,7 @@
                             </mat-error>
                         </mat-form-field>
                     </div>
-                    <div class="flex mx-2" *ngIf="this.preferences.logbookOptions.showPIC">
+                    <div class="flex mx-2" *ngIf="this.appConfig.logbookOptions.showPIC">
                         <mat-form-field class="flex-auto">
                             <mat-label>PIC</mat-label>
                             <input
@@ -388,7 +388,7 @@
                             </mat-error>
                         </mat-form-field>
                     </div>
-                    <div class="flex mx-2" *ngIf="this.preferences.logbookOptions.showSIC">
+                    <div class="flex mx-2" *ngIf="this.appConfig.logbookOptions.showSIC">
                         <mat-form-field class="flex-auto">
                             <mat-label>SIC</mat-label>
                             <input
@@ -408,7 +408,7 @@
                             </mat-error>
                         </mat-form-field>
                     </div>
-                    <div class="flex mx-2" *ngIf="this.preferences.logbookOptions.showGroundSim">
+                    <div class="flex mx-2" *ngIf="this.appConfig.logbookOptions.showGroundSim">
                         <mat-form-field class="flex-auto">
                             <mat-label>Ground Simulator</mat-label>
                             <input
