+++ src/Nisshi/ClientApp/src/app/modules/logbook/form/logbook-form.component.ts
@@ -7,8 +7,8 @@ import { AircraftService } from 'app/core/aircraft/aircraft.service';
 import { AirportService } from 'app/core/airport/airport.service';
 import { Aircraft } from 'app/core/aircraft/aircraft.types';
 import { Airport } from 'app/core/airport/airport.types';
-import { Preferences } from 'app/core/preferences/preferences.types';
-import { PreferencesService } from 'app/core/preferences/preferences.service';
+import { AppConfig } from 'app/core/config/app.config';
+import { AppConfigService } from 'app/core/config/appconfig.service';
 
 /**
  * Form that adds/edits an logbook
@@ -24,7 +24,7 @@ export class LogbookFormComponent implements OnInit
     form: FormGroup;
     aircraft: Aircraft[];
     airports: Airport[];
-    preferences: Preferences;
+    appConfig: AppConfig;
 
     /**
      * Constructor
@@ -33,7 +33,7 @@ export class LogbookFormComponent implements OnInit
                 private logbookEntryService: LogbookEntryService,
                 private aircraftService: AircraftService,
                 private airportService: AirportService,
-                private preferencesService: PreferencesService,
+                private appConfigService: AppConfigService,
                 private route: ActivatedRoute,
                 private router: Router,
                 private confirmation: ConfirmationService)
@@ -85,7 +85,7 @@ export class LogbookFormComponent implements OnInit
                 });
         }
 
-        this.preferences = this.preferencesService.preferences$;
+        this.appConfig = this.appConfigService.appConfig$;
     }
 
     // -----------------------------------------------------------------------------------------------------
